add abstract class for DownloaderLogService with default logging implementation
